app: split out Component crate into subcrate
