+++ component/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "penumbra-component"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+penumbra-storage = { path = "../storage" }
+async-trait = "0.1.52"
+anyhow = "1"
+tendermint = "0.29.0"
