+++ component/src/lib.rs
@@ -1,0 +1,3 @@
+mod component;
+
+pub use component::Component;
